refactor: rename the Eth Contract ABI wrappers and make them internal
