+++ credential-status/src/test/java/me/uport/sdk/credentialstatus/UniversalStatusResolverTests.kt
@@ -1,5 +1,3 @@
-@file:Suppress("UndocumentedPublicFunction", "UndocumentedPublicClass")
-
 package me.uport.sdk.credentialstatus
 
 import assertk.all
