+++ publishing.gradle
@@ -2,7 +2,6 @@
 // bintray_api_key
 // bintray_username
 
-
 def gitUrl = "https://github.com/uport-project/kotlin-did-jwt.git"
 def groupID = "com.github.uport-project.kotlin-did-jwt"
 def siteUrl = "https://developer.uport.me"
@@ -53,7 +52,7 @@ subprojects { subproject ->
             }
 
             task sourcesJar(type: Jar) {
-                classifier = 'sources'
+                archiveClassifier.set('sources')
                 if (subproject.hasProperty("android")) {
                     from android.sourceSets.main.javaDirectories
                 } else {
