+++ universal-did/src/main/java/me/uport/sdk/universaldid/DIDDocument.kt
@@ -1,4 +1,4 @@
-@file:Suppress("EXPERIMENTAL_API_USAGE")
+@file:Suppress("unused")
 
 package me.uport.sdk.universaldid
